fix: close TODO_BUGS §4 lax decodes + §2.7 TVAL + §3.5-3.7 device minors
Close the remaining open items in docs/todo/TODO_BUGS.md. The interpreter
(decode.c/exec_fpsimd.c) is the shared backstop: predecode.c marks any
encoding it can't specialize PD_GENERIC and both -pd and -jit fall back to
it, so most guards land in one place. Only the two encodings the fast
engines specialize are mirrored.

§4 lax decodes — UNDEF unallocated encodings consistently across all three
engines:
  - decode.c: vector ldst_register opc&2 with size!=0; add/sub extended
    imm3>4; logical + add/sub shifted-register 32-bit imm6>=32; add/sub
    shifted ROR; 3-source widening/high multiplies with sf=0.
  - exec_fpsimd.c: PMUL size!=0; simd_copy reserved imm5/imm4/Q via new
    simd_copy_enc_valid predicate.
  - predecode.c: mirror add/sub-extended imm3>4 (the only §4 case -pd
    specializes).
  - frontend.c: gate VC_COPY with a duplicated fe_simd_copy_valid so the
    JIT leaves reserved copy encodings GENERIC instead of executing them.

§2.7 — CNTP/CNTV TVAL reads now return the architectural signed-32-bit
view ((s32) not (s64)) in sysreg.c.

§3.5 — virtio-blk advertises VIRTIO_BLK_F_FLUSH (fsync handler already
existed) and no longer counts a failed read chunk into used_len.

§3.6 — fw_cfg set_item bounds against items[] overflow.

§3.7 — virtio-9p h_readdir returns EINVAL when the first entry can't fit
the client's count (an empty Rreaddir was read as spurious EOF).

Add tests/asm/m16_laxdecode.S regression test: each newly-guarded
encoding (derived from a valid neighbour, confirmed <unknown> by
llvm-objdump) must trap EC_UNKNOWN; interleaved valid neighbours must not.
Passes under the interpreter, -pd, -jit, and the a64 backend under
qemu-aarch64. Full firmware->UEFI->Linux boot reaches userspace with zero
UNDEF/UNIMPL/UNKNOWN.

+++ src/sysreg.c
@@ -105,13 +105,13 @@ static void do_mrs(CPU *c, unsigned key, unsigned Rt) {
         case KEY(3,3,14,0,0): v = c->cntfrq; break;          /* CNTFRQ_EL0 */
         case KEY(3,3,14,0,1): v = timer_count(c, false); break; /* CNTPCT_EL0 */
         case KEY(3,3,14,0,2): v = timer_count(c, true); break;  /* CNTVCT_EL0 */
-        case KEY(3,3,14,2,0): v = (u64)(s64)(c->cntp_cval - timer_count(c, false)); break; /* CNTP_TVAL */
+        case KEY(3,3,14,2,0): v = (u64)(s32)(c->cntp_cval - timer_count(c, false)); break; /* CNTP_TVAL (signed 32-bit view) */
         case KEY(3,3,14,2,1):                                /* CNTP_CTL_EL0 (+ISTATUS) */
             v = c->cntp_ctl & 3;
             if ((c->cntp_ctl & 1) && timer_count(c, false) >= c->cntp_cval) v |= 4;
             break;
         case KEY(3,3,14,2,2): v = c->cntp_cval; break;       /* CNTP_CVAL_EL0 */
-        case KEY(3,3,14,3,0): v = (u64)(s64)(c->cntv_cval - timer_count(c, true)); break;  /* CNTV_TVAL */
+        case KEY(3,3,14,3,0): v = (u64)(s32)(c->cntv_cval - timer_count(c, true)); break;  /* CNTV_TVAL (signed 32-bit view) */
         case KEY(3,3,14,3,1):                                /* CNTV_CTL_EL0 (+ISTATUS) */
             v = c->cntv_ctl & 3;
             if ((c->cntv_ctl & 1) && timer_count(c, true) >= c->cntv_cval) v |= 4;
